Optimize the Transformation class
This commit does not change the semantics of any method.
This commit only applies the changes to the latest demo c21-p1.

Profiling the application showed a big overhead for JOML methods, especially the rotate*() methods, which call java.lang.Math.sin/cos. These changes optimize the use of JOML to the point where no JOML method is within 0.01% of the application's CPU usage anymore:
- use fast Matrix4f.translationRotateScale() to build a model transformation (this is a dedicated method to build the common T * R * S transformation)
- use Quaternionf rather than three Euler angles to avoid calling sin/cos many times per frame when building the model transformation for each GameItem

Others:
- small typo
- refactor Transformation to reuse buildModelMatrix() in various places